feat: Add npmrc support check for Bun package manager (fixes: #93)
This commit introduces a check to determine if the Bun package manager
supports npmrc. If the version of Bun is 1.1.18 or higher, it is
considered to support npmrc. This information is used to decide whether
to create a bunfig.toml file or not. The tests have been updated
accordingly to check for either bunfig.toml or .npmrc based on the
version of Bun.